requirements.txt
python-telegram-bot==20.7
requests==2.31.0
Pillow==10.2.0
qrcode[pil]==7.4.2
pytesseract==0.3.10
APScheduler==3.10.4
python-dotenv==1.0.0
aiohttp==3.9.3
aiofiles==23.2.1
pytz==2024.1
